requirements.txt
Flask
twilio
openai
requests
python-dotenv
ffmpeg-python
